dbem_get_program_link no longer returns home page link
If there wasn't a program assigned to an event, dbem_get_program_link
would use the home page, as returned by get_page_link. Now it checks to
see if there's a program assigned before doing its thing, returning an
empty string if no program is assigned.

Also tucked in a couple of <?php tags to substitute for the broken <?
tags in the original file. Sloppy, sloppy.