dbem_get_program_link now returns message for empty program
No easy way to do an if statement to make a conditional line break in
the #_PROGRAM tag, so if we're always going to have something there, it
might as well be the informative "Program to be announced" message.
Added bonus: no need to manually write that out if an event has no
program assigned to it.