Change from custom program text to custom program title
Mostly a conceptual change in how it works. Instead of putting the full
description of a custom program into a textarea, put just the title of
the custom program into a text input field. This distinction lets the
custom text easily replace the existing title/link used for an attached
program, which appears in the short text of the sidebar on the home
page. A full description may be included in the existing description
field for the event.

Also added some calls to stripslashes to prevent continuous recursive
insertion of backslashes into text with single quotes in it. Silly
fucking PHP.